requirements.txt
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
thop>=0.1.1
umap-learn>=0.5.3
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=9.5.0
timm>=0.9.0
torchao>=0.1.0
